Fix for force view mode (#836)
* Fix for force view mode

The reaosn for the change:
- Since the default view mode value for all the view types in the Elementum settings are equal to "0",
Each navigation to any place in the addon force the view mode to be changed to "List" view (In my skin).
--- Probably "Container.SetViewMode(0)" in Kodi 19.2 force it to be changed to "List" as default since there is no view mode id of "0" in my skin (And I guess in others too)
- I also have another issue regarding to this:
One of the Lists in Elementum is used as a widget in my skin. 
So on each refreshing of the widget (Kodi startup/Reload skin etc...) actually runs Elemetum which activate the force view mode changing.
It cause that after each widget refreshing the first window in kodi that you navigate to it, its view mode will be changed according to what set in Elementum.

The fix:
It's so simple.
If the view mode is equal to "0" (Which is the defaul value), don't do any view mode changing.
Otherwise it will change it according to the Elementum settings.
In this case,  every change of the view modes inside Elementum by the skin context menu, the view mode is kept in Kodi/Skin/Addon cache or something (So at least I don't need this "Force" view mode changing). If for others these view mode changes are not kept in memory, they will need these options of "Force" changing and use it as always.

It was tested and works well.

Thanks.

* Fix for force view mode - Condition fix